requirements.txt
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
google_api_python_client==1.7.4
GoogleAppEngineCloudStorageClient==1.9.22.1
google_auth==1.5.1
google_auth_httplib2==0.0.3
gspread==3.0.1
httplib2==0.11.3
idna==2.7
oauth2client==4.1.2
pyasn1==0.4.4
pyasn1_modules==0.2.2
requests==2.19.1
requests_toolbelt==0.8.0
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
